requirements.txt
streamlit
pandas
sentence-transformers
python-dotenv
requests
torch
torchvision
fastapi
uvicorn
sentence-transformers
torch
pandas
